rafactor: types
